fix(tests)

+++ src/interfaces/IBridge.sol
@@ -9,5 +9,5 @@ interface IBridge {
     /// @dev Calls this method on the L1 Proxy contract to relay roots and timestamps to WorldID supported chains.
     /// @param root The latest Semaphore root.
     /// @param opGasLimit The gas limit for the Optimism transaction (how much gas to buy on Optimism with the message)
-    function sendRootMultichain(uint256 root, uint256 opGasLimit) external;
+    function sendRootMultichain(uint256 root, uint32 opGasLimit) external;
 }
